% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{clean_logs}
\alias{clean_logs}
\title{Clean FAH Log Data}
\usage{
clean_logs(logs_df)
}
\arguments{
\item{logs_df}{The log files tibble as read in by read_fah_logs()}
}
\value{
\strong{clean_logs:} Tibble of cleaned log data.
}
\description{
Cleans the output of \code{read_fah_logs()}
}
\examples{
# Windows 10:
my_log_data <- read_fah_logs("~/../AppData/Roaming/FAHClient/logs/")
my_clean_log_data <- clean_logs(my_log_data)
summary(my_clean_log_data)
}
