% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{fah_web_palette}
\alias{fah_web_palette}
\title{FAH Website Palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
fah_web_palette
}
\description{
FAH Website Palette
}
\keyword{datasets}
