% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_latest_work_unit_details}
\alias{get_latest_work_unit_details}
\title{Get Latest Complete Work Unit Details}
\usage{
get_latest_work_unit_details(work_units_df)
}
\description{
Get Latest Complete Work Unit Details
}
