% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_processing_time_summary}
\alias{get_processing_time_summary}
\title{Get Processing Time Summary}
\usage{
get_processing_time_summary(parsed_log)
}
\arguments{
\item{parsed_log}{A tibble of FAH Client logs that are parsed
(\code{read_fah_logs()}) and cleaned (\code{clean_logs()}).}
}
\value{
Tibble summary of processing time for different work units by
folding slot.
}
\description{
Get Processing Time Summary
}
\examples{
read_fah_logs("~/../AppData/Roaming/FAHClient/logs/") \%>\%
  clean_logs() \%>\%
  get_processing_time_summary()
}
