% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_credits}
\alias{plot_credits}
\title{Plot Credits by Folding Slot}
\usage{
plot_credits(credits_df, all_slots = FALSE)
}
\arguments{
\item{parsed_log}{A tibble of FAH Client logs that are parsed
(\code{read_fah_logs()}) and cleaned (\code{clean_logs()}).}
}
\value{
A plot of credits acquired per day.
}
\description{
Given the output of \code{get_credits()}, this will create a stacked bar chart of
credits awarded by folding slot. Set \code{all_slots = TRUE} if you want to see
total credits of all folding slots.
}
\examples{
read_fah_logs("~/../AppData/Roaming/FAHClient/logs/") \%>\%
  clean_logs() \%>\%
  get_credits() \%>\%
  plt_credits()
}
