% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cumulative_network_usage}
\alias{plot_cumulative_network_usage}
\title{Plot Cumulative Network Usage}
\usage{
plot_cumulative_network_usage(network_usage_daily_summary)
}
\description{
Plot Cumulative Network Usage
}
